Bug: User should be able to comment and reply on article comments - ensure to add tests [Bug #162730416]
